+++ Civi/DataProcessor/Source/Event/ParticipantSource.php
@@ -6,12 +6,50 @@
 
 namespace Civi\DataProcessor\Source\Event;
 
+use Civi\DataProcessor\DataFlow\MultipleDataFlows\DataFlowDescription;
+use Civi\DataProcessor\DataFlow\MultipleDataFlows\SimpleJoin;
+use Civi\DataProcessor\DataFlow\SqlTableDataFlow;
+use Civi\DataProcessor\DataSpecification\DataSpecification;
+use Civi\DataProcessor\DataSpecification\FieldSpecification;
+use Civi\DataProcessor\DataSpecification\Utils as DataSpecificationUtils;
+use Civi\DataProcessor\ProcessorType\AbstractProcessorType;
 use Civi\DataProcessor\Source\AbstractCivicrmEntitySource;
 
 use Civi\DataProcessor\Utils\AlterExportInterface;
 use CRM_Dataprocessor_ExtensionUtil as E;
 
 class ParticipantSource extends AbstractCivicrmEntitySource implements AlterExportInterface {
+
+  /**
+   * @var SqlTableDataFlow
+   */
+  protected $participantPaymentDataFlow;
+
+  /**
+   * @var SimpleJoin
+   */
+  protected $participantPaymentJoin;
+
+  public function __construct() {
+    parent::__construct();
+  }
+
+  public function initialize() {
+    $this->initializeParticipantPayment();
+    parent::initialize();
+  }
+
+  protected function initializeParticipantPayment() {
+    if (!$this->participantPaymentDataFlow) {
+      $this->participantPaymentDataFlow = new SqlTableDataFlow('civicrm_participant_payment', $this->getSourceName().'_participant_payment');
+      DataSpecificationUtils::addDAOFieldsToDataSpecification('CRM_Event_DAO_ParticipantPayment', $this->participantPaymentDataFlow->getDataSpecification(), array('id'), '', 'participant_payment_', E::ts('Participant Payment :: '));
+    }
+    if (!isset($this->additionalDataFlowDescriptions['civicrm_participant_payment'])) {
+      $this->participantPaymentJoin = new SimpleJoin($this->getSourceName(), 'id', $this->participantPaymentDataFlow->getName(), 'participant_id', 'LEFT');
+      $this->participantPaymentJoin->setDataProcessor($this->dataProcessor);
+      $this->additionalDataFlowDescriptions['civicrm_participant_payment'] = new DataFlowDescription($this->participantPaymentDataFlow, $this->participantPaymentJoin);
+    }
+  }
 
   /**
    * Returns the entity name
@@ -29,6 +67,33 @@ class ParticipantSource extends AbstractCivicrmEntitySource implements AlterExpo
    */
   protected function getTable() {
     return 'civicrm_participant';
+  }
+
+  /**
+   * Load the fields from this entity.
+   *
+   * @param DataSpecification $dataSpecification
+   * @throws \Civi\DataProcessor\DataSpecification\FieldExistsException
+   */
+  protected function loadFields(DataSpecification $dataSpecification, $fieldsToSkip=array()) {
+    parent::loadFields($dataSpecification, $fieldsToSkip);
+    DataSpecificationUtils::addDAOFieldsToDataSpecification('CRM_Event_DAO_ParticipantPayment', $dataSpecification, array('id', 'participant_id'), '', 'participant_payment_', E::ts('Participant Payment :: '));
+  }
+
+  /**
+   * Ensure that filter or aggregate field is accesible in the query
+   *
+   * @param FieldSpecification $field
+   *
+   * @return \Civi\DataProcessor\DataFlow\AbstractDataFlow|null
+   * @throws \Exception
+   */
+  public function ensureField(FieldSpecification $field) {
+    if ($field->alias == 'participant_payment_contribution_id') {
+      $this->initializeParticipantPayment();
+      return $this->participantPaymentDataFlow;
+    }
+    return parent::ensureField($field);
   }
 
   /**